@@ -21,8 +21,8 @@
     }
   },
   "require": {
-    "php": "^7.2",
-    "laravel/framework": "^6||^7",
-    "symfony/process":"~4.0||^5"
+    "php": "^7.4",
+    "laravel/framework": "^7",
+    "symfony/process":"^5"
   }
 }
